+++ crates/ronami-core/src/net.rs
@@ -71,8 +71,8 @@ pub fn client_from_env() -> reqwest::Client {
 /// [issue 223]: https://github.com/teloxide/teloxide/issues/223
 pub fn default_reqwest_settings() -> reqwest::ClientBuilder {
     reqwest::Client::builder()
-        .connect_timeout(Duration::from_secs(5))
-        .timeout(Duration::from_secs(17))
+        .connect_timeout(Duration::from_secs(10))
+        .timeout(Duration::from_secs(60))
         .tcp_nodelay(true)
 }