+++ crates/ronami-core/src/prelude.rs
@@ -3,6 +3,7 @@
 #[doc(no_inline)]
 pub use crate::{
     payloads::setters::*,
+    requester_forward,
     requests::{Request, Requester, RequesterExt},
     types::{ChatId, UserId},
     Bot,